Use unsigned for frames and bytes in FifoBuffer
FifoBuffer::convertFramesToBytes() uses signed types, and uses negative
values as a form of error checking.

This CL updates FifoBuffer to used uint32_t when representing frames and
bytes, and changes the error checking to rely on comparisons rather than
negative values.

+++ src/opensles/AudioStreamBuffered.cpp
@@ -134,7 +134,7 @@ ResultWithValue<int32_t> AudioStreamBuffered::transfer(void *buffer,
             result = mFifoBuffer->write(data, framesLeft);
         }
         if (result > 0) {
-            data += mFifoBuffer->convertFramesToBytes(result);
+            data += mFifoBuffer->convertFramesToBytes(static_cast<uint32_t>(result));
             framesLeft -= result;
         }
